requirements.txt
discord.py
python-dotenv  
aiosqlite  
psutil  
asyncio  
aiohttp  
colorlog 